requirements.txt
streamlit
python-dotenv
pdfplumber
python-docx
pandas
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
